Docs for interpreting "Unsatisfiable requirement"
